pyproject.toml
[project]
name = "dreamer"
version = "0.1.0"
description = "SonicVision Studio CLI - Audio to Visual Storyboard"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.56.0",
    "hatchling>=1.28.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.1",
]

[project.scripts]
dreamer = "dreamer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
